RIOTCore: Fix a number of issues * Read/Write addresses using wrong base * Connect read of SWCHA/SWCHB * Fix timer prescale values, and using wrong storage type * Fix behaviour on timer underflow

+++ src/RIOTCore.h
@@ -24,8 +24,8 @@ public:
   void ClearSWCHB(uint8_t val);
 
 private:
-  uint8_t m_TimerPrescale;
-  uint8_t m_TimerPrescaleCnt;
+  uint16_t m_TimerPrescale;
+  uint16_t m_TimerPrescaleCnt;
   uint8_t m_SWCHA;
   uint8_t m_SWCHB;
   uint8_t m_SWACNT;
